+++ src/main/java/domain/Person.java
@@ -1,4 +1,4 @@
-package persistence.example;
+package domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
